renamed graphql -> graphqlQueries
